backend/requirements.txt
fastapi
uvicorn
python-multipart
groq
python-dotenv
pydantic
PyMuPDF
python-docx
slowapi
